# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "api_h2o_local",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://34.123.11.193:3000",
      "variables": {}
    }
  ],
  "paths": {
...
    "/typeUser": {
      "post": {
        "tags": [
          "AdministrarTiposUsuarios"
        ],
        "summary": "createTypeUser",
        "operationId": "createTypeUser",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/createTypeUserRequest"
                  },
                  {
                    "example": {
                      "name": "cliente mayoreo"
                    }
                  }
                ]
              },
              "example": {
                "name": "cliente mayoreo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/User/5": {
      "put": {
        "tags": [
          "AdministracionDeUsuarios"
        ],
        "summary": "updateUser",
        "operationId": "updateUser",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/updateUserRequest"
                  },
                  {
                    "example": {
                      "name": "cliente02",
                      "password": "admin",
                      "type": 5,
                      "status": 1
                    }
                  }
                ]
              },
              "example": {
                "name": "cliente02",
                "password": "admin",
                "type": 5,
                "status": 1
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/user": {
      "post": {
        "tags": [
          "AdministracionDeUsuarios"
        ],
        "summary": "createUser",
        "operationId": "createUser",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/createUserRequest"
                  },
                  {
                    "example": {
                      "name": "cliente03",
                      "password": "admin",
                      "type": 6
                    }
                  }
                ]
              },
              "example": {
                "name": "cliente03",
                "password": "admin",
                "type": 6
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
      "createTypeUserRequest": {
        "title": "createTypeUserRequest",
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "example": {
          "name": "cliente mayoreo"
        }
      },
...
      "updateUserRequest": {
        "title": "updateUserRequest",
        "required": [
          "name",
          "password",
          "type",
          "status"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "type": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "name": "cliente02",
          "password": "admin",
          "type": 5,
          "status": 1
        }
      },
      "createUserRequest": {
        "title": "createUserRequest",
        "required": [
          "name",
          "password",
          "type"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "type": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "name": "cliente03",
          "password": "admin",
          "type": 6
        }
      },
...
    }
  },
...
}